improved build process

+++ router/context_database.go
@@ -47,7 +47,7 @@ func CastFieldsPg(model *models.Model, m map[string]any) {
 		} else if field.Flavor == "bigint" && m[field.Name] != nil {
 			m[field.Name] = m[field.Name].(int64)
 		} else if field.Flavor == "smallint" && m[field.Name] != nil {
-			m[field.Name] = m[field.Name].(int16)
+			m[field.Name] = m[field.Name].(int64)
 		} else if field.Flavor == "float" && m[field.Name] != nil {
 			m[field.Name] = m[field.Name].(float64)
 		} else if field.Flavor == "bool" && m[field.Name] != nil {
